+++ scripts/opt/raspberrypi-ua-netinst/install.sh
@@ -660,6 +660,7 @@ case "${rpi_hardware}" in
 	"a22082") rpi_hardware_version="3 Model B" ;;
 	"a32082") rpi_hardware_version="3 Model B" ;;
 	"a020d3") rpi_hardware_version="3 Model B+" ;;
+	"9020e0") rpi_hardware_version="3 Model A+" ;;
 	*) rpi_hardware_version="unknown (${rpi_hardware})" ;;
 esac
 
@@ -673,6 +674,10 @@ echo "Running on Raspberry Pi version ${rpi_hardware_version}"
 echo "=================================================="
 echo "https://github.com/FooDeas/raspberrypi-ua-netinst/"
 echo "=================================================="
+
+echo -n "Remounting TempFS... "
+mount -o remount,size=80% / || fail
+echo "OK"
 
 echo -n "Mounting boot partition... "
 mount "${bootpartition}" /boot || fail
@@ -865,6 +870,23 @@ fi
 
 echo "  online_config = ${online_config}"
 echo
+
+# create symlink for other kernel modules if needed
+if [ ! -e "/lib/modules/$(uname -r)" ]; then
+	echo "Kernel modules for the kernel version \"$(uname -r)\" could not be found. Searching for alternatives..."
+	if [[ "$(uname -r)" =~ -v7\+$ ]]; then
+		kernel_modulepath="$(find /lib/modules/ -maxdepth 1 -type d ! -path /lib/modules/ | grep -e "[^/]\+-v7+$" | head -1)"
+	else
+		kernel_modulepath="$(find /lib/modules/ -maxdepth 1 -type d ! -path /lib/modules/ | grep -ve "[^/]\+-v7+$" | head -1)"
+	fi
+	if [ -z "${kernel_modulepath}" ] ; then
+		echo "ERROR: No kernel modules could be found!"
+		fail_blocking
+	fi
+	#kernel_modulename="$("${kernel_modulepath}" | grep -oe "[^/]\+$"\")"
+	echo "  Using modules of kernel version \"$(echo "${kernel_modulepath}" | grep -oe "[^/]\+$")\"."
+	ln -s "${kernel_modulepath}" "/lib/modules/$(uname -r)"
+fi
 
 # depmod needs to update modules.dep before using modprobe
 depmod -a
@@ -1119,7 +1141,7 @@ if [ -z "${cdebootstrap_cmdline}" ]; then
 
 	# always add packages if requested or needed
 	if [ "${firmware_packages}" = "1" ]; then
-		custom_packages_postinstall="${custom_packages_postinstall},firmware-atheros,firmware-brcm80211,firmware-libertas,firmware-ralink,firmware-realtek"
+		custom_packages_postinstall="${custom_packages_postinstall},firmware-atheros,firmware-brcm80211,firmware-libertas,firmware-misc-nonfree,firmware-realtek"
 	fi
 	if [ -n "${locales}" ] || [ -n "${system_default_locale}" ]; then
 		custom_packages="${custom_packages},locales"
@@ -1155,7 +1177,7 @@ if [ -z "${cdebootstrap_cmdline}" ]; then
 		base_packages_postinstall="${base_packages_postinstall},raspberrypi-kernel"
 	fi
 	base_packages_postinstall="${custom_packages_postinstall},${base_packages_postinstall}"
-	
+
 	# minimal
 	minimal_packages="cpufrequtils,ifupdown,net-tools,openssh-server,dosfstools"
 	if [ "${init_system}" != "systemd" ]; then
@@ -1567,11 +1589,10 @@ for i in $(seq 1 "${installer_pkg_downloadretries}"); do
 	else
 		unset http_proxy
 		if [ "${i}" -eq "${installer_pkg_downloadretries}" ]; then
-			echo
-			echo "  ERROR: ${cdebootstrap_exitcode}"
+			echo -e "\n  ERROR: ${cdebootstrap_exitcode}"
 			fail
 		else
-			echo "  ERROR: ${cdebootstrap_exitcode}, trying again ($((i+1))/${installer_pkg_downloadretries})..."
+			echo -e "\n  ERROR: ${cdebootstrap_exitcode}, trying again ($((i+1))/${installer_pkg_downloadretries})..."
 		fi
 	fi
 done
@@ -1763,7 +1784,7 @@ if echo "${cdebootstrap_cmdline} ${packages_postinstall}" | grep -q "ifupdown"; 
 		echo "auto lo" >> /rootfs/etc/network/interfaces
 		echo "iface lo inet loopback" >> /rootfs/etc/network/interfaces
 	fi
-	
+
 	# configured interface
 	echo >> /rootfs/etc/network/interfaces
 	echo "allow-hotplug ${ifname}" >> /rootfs/etc/network/interfaces
@@ -1778,7 +1799,7 @@ if echo "${cdebootstrap_cmdline} ${packages_postinstall}" | grep -q "ifupdown"; 
 			echo "    gateway ${ip_gateway}"
 		} >> /rootfs/etc/network/interfaces
 	fi
-	
+
 	# wlan config
 	if echo "${ifname}" | grep -q "wlan"; then
 		if [ -e "${wlan_configfile}" ]; then
@@ -1794,7 +1815,7 @@ if echo "${cdebootstrap_cmdline} ${packages_postinstall}" | grep -q "ifupdown"; 
 			echo "iface eth0 inet dhcp"
 		} >> /rootfs/etc/network/interfaces
 	fi
-	
+
 	# Customize cmdline.txt
 	if [ "${disable_predictable_nin}" = "1" ]; then
 		# as described here: https://www.freedesktop.org/wiki/Software/systemd/PredictableNetworkInterfaceNames
@@ -1802,7 +1823,7 @@ if echo "${cdebootstrap_cmdline} ${packages_postinstall}" | grep -q "ifupdown"; 
 		line_add cmdline_custom "net.ifnames=0"
 		ln -s /dev/null /rootfs/etc/udev/rules.d/75-persistent-net-generator.rules
 	fi
-	
+
 	# copy resolv.conf
 	echo -n "  Configuring nameserver... "
 	if [ -e "/etc/resolv.conf" ]; then
@@ -1816,7 +1837,7 @@ if echo "${cdebootstrap_cmdline} ${packages_postinstall}" | grep -q "ifupdown"; 
 		echo "MISSING !"
 		fail
 	fi
-	
+
 	echo "OK"
 fi
 
@@ -2096,7 +2117,7 @@ if [ "${kernel_module}" = true ]; then
 	else
 		echo "FAILED !"
 	fi
-	
+
 	unset DEBIAN_FRONTEND
 fi